Finalizando Cadastro aluno

+++ src/view/CadastroAlunos.java
@@ -1,7 +1,5 @@
 package view;
 
-import java.awt.event.ActionEvent;
-import java.awt.event.ActionListener;
 import java.awt.event.FocusAdapter;
 import java.awt.event.FocusEvent;
 import java.sql.SQLException;
@@ -9,8 +7,6 @@ import java.text.ParseException;
 import java.text.SimpleDateFormat;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
-
-import javax.swing.AbstractAction;
 import javax.swing.JComboBox;
 import javax.swing.JFormattedTextField;
 import javax.swing.JLabel;
@@ -18,12 +14,12 @@ import javax.swing.JScrollPane;
 import javax.swing.JTextArea;
 import javax.swing.JTextField;
 import javax.swing.ScrollPaneConstants;
-import javax.swing.SwingConstants;
 import javax.swing.text.MaskFormatter;
 
 import dao.AlunoDAO;
 import model.AlunoModel;
 
+@SuppressWarnings("serial")
 public class CadastroAlunos extends MasterDialogCad {
 
 	private JLabel Aluno, DataNasc, Tel, Email, Obs, Endereco, Complemento, Bairro, Estado, CEP, Numero, Cidade, Pais,
@@ -35,24 +31,26 @@ public class CadastroAlunos extends MasterDialogCad {
 	private JComboBox<String> ComboSexo;
 	private AlunoDAO alunoDao;
 	private AlunoModel aluno, alunoChange;
-	
-	
+	private BuscarAluno busca;
+
+
 
 	public CadastroAlunos() throws ParseException {
 
+		alunoDao = new AlunoDAO();
+		
 		setSize(690, 538);
 		setTitle("Cadastro de Alunos");
 		setLayout(null);
 		setResizable(false);
 		setClosable(true);
-		setDefaultCloseOperation(DISPOSE_ON_CLOSE);
+		setDefaultCloseOperation(DISPOSE_ON_CLOSE);		
 		setVisible(true);
-		
-		
+
 	}
-	
+
 	protected boolean actionDelete() {
-		if(aluno!=null) {
+		if((aluno!=null) && (!isInserting)) {
 			try {
 				alunoDao.deleteAluno(aluno.getcodigoAluno());
 				return true;
@@ -63,19 +61,65 @@ public class CadastroAlunos extends MasterDialogCad {
 			return false;
 		}
 	}
-	
+
 	protected boolean actionAdd() {
-		try {
-			aluno = new AlunoModel();
-			fillFields();
-			return true;
-		} catch (Exception e) {
+		if(!isInserting) {
+			try {
+				aluno = new AlunoModel();
+				return true;
+			} catch (Exception e) {
+				return false;
+			}			
+		}else {
 			return false;
-		}			
+		}
 	}
-	
-	private void fillFields() {
-		
+
+	protected void actionSearch() {
+		busca = new BuscarAluno();
+		try {						
+			busca.addWindowListener(eventWindowSearchClosed);				
+		} catch (Exception e2) {
+			busca = null;
+		}
+	}
+
+	protected boolean afterSearch() {
+		if(busca.alunoReturn!=null) {
+			aluno = busca.alunoReturn;
+			return true;
+		}
+		return false;
+	}
+
+	protected boolean actionSave() {
+		if(alunoChange!=null) {
+			try {
+				if(isInserting) {
+					alunoDao.createAluno(alunoChange);
+				}else {
+					alunoDao.updateAluno(alunoChange);
+				}
+				return true;
+			}catch (Exception e) {
+				return false;
+			}
+		}else {
+			return false;
+		}
+	}
+
+	protected boolean actionCancel() {
+		try {			
+			aluno = null;
+			alunoChange = null;			
+		}catch (Exception e) {
+		}				
+		return true;		
+	}
+
+	protected void fillFields() {
+
 		JTAluno.setText(aluno.getAluno());
 		JTDataNasc.setText(new SimpleDateFormat("dd-mm-yyyy").format(aluno.getDataNascimento()));
 		JTTel.setText(aluno.getTelefone());
@@ -92,194 +136,203 @@ public class CadastroAlunos extends MasterDialogCad {
 		JTObs.setText(aluno.getObservacao());
 		JTCel.setText(aluno.getCelular());	
 		ComboSexo.setSelectedIndex((aluno.getSexo() == 'M') ? 1 : 2);
-		
+
 		alunoChange = aluno;
-		
+
 	}
-	
-	protected void setFieldsEnabled(boolean enabled) {
-		utils.setFieldsEnabled(getContentPane(), enabled);
-	}
-	
+
 	protected void subComponnents() {
-		
-		btnSearch.addActionListener(new ActionListener() {
 
-			public void actionPerformed(ActionEvent e) {
 
-				BuscarAluno busca = new BuscarAluno();
-				aluno = busca.alunoReturn;
-				fillFields();
 
-			}
-		});
-		btnSearch.setHorizontalAlignment(SwingConstants.LEFT);
-		btnSearch.setHorizontalTextPosition(SwingConstants.RIGHT);
-		btnSearch.setBounds(10, 10, 120, 35);
-		getContentPane().add(btnSearch);		
-	
-		
+
 		JTAluno = new JTextField();
 		JTAluno.addFocusListener(new FocusAdapter() {
-			
+
 			public void focusLost(FocusEvent e) {
-				alunoChange.setAluno(JTAluno.getText());
+				if((!e.isTemporary()) && (alunoChange!=null)) {
+					alunoChange.setAluno(JTAluno.getText());
+				}
 			}
-			
+
 		});
 		JTAluno.setBounds(135, 50, 250, 26);
 		getContentPane().add(JTAluno);
-		
+
 		try {
 			JTDataNasc = new JFormattedTextField(new MaskFormatter("##-##-####"));
 		} catch (ParseException e2) {
 		}
 		JTDataNasc.setText(DateTimeFormatter.ofPattern("dd-MM-yyyy").format(LocalDate.now()));
 		JTDataNasc.addFocusListener(new FocusAdapter() {
-			
+
 			public void focusLost(FocusEvent e) {
-/*				try {
+				try {
 					if(!JTDataNasc.hasFocus()) {
 						alunoChange.setDataNascimento(new SimpleDateFormat("dd-mm-yyyy").parse(JTDataNasc.getText()));
-						JOptionPane.showMessageDialog(null,"Saiu");
+						//JOptionPane.showMessageDialog(null,"Saiu");
 					}
 				} catch (ParseException e1) {
 					e1.printStackTrace();
 				}
-*/			}
-			
+			}
+
 		});
 		JTDataNasc.setBounds(135, 80, 250, 26);
 		getContentPane().add(JTDataNasc);
 
 		JTTel = new JTextField();
 		JTTel.addFocusListener(new FocusAdapter() {
-			
+
 			public void focusLost(FocusEvent e) {
-				alunoChange.setTelefone(JTTel.getText());
+				if(!e.isTemporary()) {
+					alunoChange.setTelefone(JTTel.getText());
+				}
 			}
-			
+
 		});
 		JTTel.setBounds(135, 110, 250, 26);
 		getContentPane().add(JTTel);
 
 		JTEmail = new JTextField();
 		JTEmail.addFocusListener(new FocusAdapter() {
-			
+
 			public void focusLost(FocusEvent e) {
-				alunoChange.setEmail(JTEmail.getText());
+				if(!e.isTemporary()) {
+					alunoChange.setEmail(JTEmail.getText());
+				}
 			}
-			
+
 		});
 		JTEmail.setBounds(135, 140, 250, 26);
 		getContentPane().add(JTEmail);
 
 		JTEndereco = new JTextField();
 		JTEndereco.addFocusListener(new FocusAdapter() {
-			
+
 			public void focusLost(FocusEvent e) {
-				alunoChange.setEndereco(JTEndereco.getText());
+				if(!e.isTemporary()) {
+					alunoChange.setEndereco(JTEndereco.getText());
+				}
 			}
-			
+
 		});
 		JTEndereco.setBounds(100, 315, 250, 26);
 		getContentPane().add(JTEndereco);
 
 		JTComplemento = new JTextField();
 		JTComplemento.addFocusListener(new FocusAdapter() {
-			
+
 			public void focusLost(FocusEvent e) {
-				alunoChange.setComplemento(JTComplemento.getText());
+				if(!e.isTemporary()) {
+					alunoChange.setComplemento(JTComplemento.getText());
+				}
 			}
-			
+
 		});
 		JTComplemento.setBounds(100, 345, 570, 26);
 		getContentPane().add(JTComplemento);
 
 		JTBairro = new JTextField();
 		JTBairro.addFocusListener(new FocusAdapter() {
-			
+
 			public void focusLost(FocusEvent e) {
-				alunoChange.setBairro(JTBairro.getText());
+				if(!e.isTemporary()) {
+					alunoChange.setBairro(JTBairro.getText());
+				}
 			}
-			
+
 		});
 		JTBairro.setBounds(100, 375, 250, 26);
 		getContentPane().add(JTBairro);
 
 		JTEstado = new JTextField();
 		JTEstado.addFocusListener(new FocusAdapter() {
-			
+
 			public void focusLost(FocusEvent e) {
-				alunoChange.setEstado(JTEstado.getText());
+				if(!e.isTemporary()) {
+					alunoChange.setEstado(JTEstado.getText());
+				}
 			}
-			
+
 		});
 		JTEstado.setBounds(100, 405, 250, 26);
 		getContentPane().add(JTEstado);
 
 		JTCEP = new JTextField();
 		JTCEP.addFocusListener(new FocusAdapter() {
-			
+
 			public void focusLost(FocusEvent e) {
-				alunoChange.setCep(JTCEP.getText());
+				if(!e.isTemporary()) {
+					alunoChange.setCep(JTCEP.getText());
+				}
 			}
-			
+
 		});
 		JTCEP.setBounds(100, 435, 250, 26);
 		getContentPane().add(JTCEP);
 
 		JTNumero = new JTextField();
 		JTNumero.addFocusListener(new FocusAdapter() {
-			
+
 			public void focusLost(FocusEvent e) {
-				alunoChange.setNumero(JTNumero.getText());
+				if(!e.isTemporary()) {
+					alunoChange.setNumero(JTNumero.getText());
+				}
 			}
-			
+
 		});
 		JTNumero.setBounds(420, 315, 250, 26);
 		getContentPane().add(JTNumero);
 
 		JTCidade = new JTextField();
 		JTCidade.addFocusListener(new FocusAdapter() {
-			
+
 			public void focusLost(FocusEvent e) {
-				alunoChange.setCidade(JTCidade.getText());
+				if(!e.isTemporary()) {
+					alunoChange.setCidade(JTCidade.getText());
+				}
 			}
-			
+
 		});
 		JTCidade.setBounds(420, 375, 250, 26);
 		getContentPane().add(JTCidade);
 
 		JTPais = new JTextField();
 		JTPais.addFocusListener(new FocusAdapter() {
-			
+
 			public void focusLost(FocusEvent e) {
-				alunoChange.setPais(JTPais.getText());
+				if(!e.isTemporary()) {
+					alunoChange.setPais(JTPais.getText());
+				}
 			}
-			
+
 		});
 		JTPais.setBounds(420, 405, 250, 26);
 		getContentPane().add(JTPais);
 
 		JTCel = new JTextField();
 		JTCel.addFocusListener(new FocusAdapter() {
-			
+
 			public void focusLost(FocusEvent e) {
-				alunoChange.setCelular(JTCel.getText());
+				if(!e.isTemporary()) {
+					alunoChange.setCelular(JTCel.getText());
+				}
 			}
-			
+
 		});
 		JTCel.setBounds(435, 110, 235, 26);
 		getContentPane().add(JTCel);
 
 		JTObs = new JTextArea();
 		JTObs.addFocusListener(new FocusAdapter() {
-			
+
 			public void focusLost(FocusEvent e) {
-				alunoChange.setObservacao(JTObs.getText());
+				if((!e.isTemporary()) && (alunoChange!=null)) {
+					alunoChange.setObservacao(JTObs.getText());
+				}
 			}
-			
+
 		});
 		getContentPane().add(JTObs);
 		JTObs.setLineWrap(true);
@@ -291,7 +344,7 @@ public class CadastroAlunos extends MasterDialogCad {
 
 		ComboSexo = new JComboBox<String>();
 		ComboSexo.addFocusListener(new FocusAdapter() {
-			
+
 			public void focusLost(FocusEvent e) {
 				if(ComboSexo.getSelectedIndex()>0) {
 					alunoChange.setSexo((ComboSexo.getSelectedIndex() == 1) ? 'M' : 'F');
@@ -299,7 +352,7 @@ public class CadastroAlunos extends MasterDialogCad {
 					alunoChange.setSexo(' ');
 				}
 			}
-			
+
 		});
 		ComboSexo.addItem("--Selecione--");
 		ComboSexo.addItem("Masculino");
@@ -307,31 +360,8 @@ public class CadastroAlunos extends MasterDialogCad {
 		ComboSexo.setBounds(435, 80, 235, 26);
 		getContentPane().add(ComboSexo);
 
-
-		btnAdd.addActionListener(new AbstractAction() {
-			
-			@Override
-			public void actionPerformed(ActionEvent e) {
-				//A função precisa ser mudada para aceitar um model para dar insert em todos campos de uma vez
-				//alunoDao.createAluno(alunoChange);								
-				
-			}
-		});
-
-		btnSave.addActionListener(new AbstractAction() {
-			
-			@Override
-			public void actionPerformed(ActionEvent arg0) {
-				try {
-					alunoDao.updateAluno(alunoChange);
-				} catch (SQLException e) {
-				}				
-			}
-		});
-		
-		
 		//labels
-		
+
 		Aluno = new JLabel("Aluno:");		
 		Aluno.setBounds(10, 10, 100, 100);
 		getContentPane().add(Aluno);
@@ -392,7 +422,7 @@ public class CadastroAlunos extends MasterDialogCad {
 		Obs.setBounds(10, 130, 110, 100);
 		getContentPane().add(Obs);
 
+		childContainer = getContentPane();
 
-		
 	}
 }
